Fix HandleInput instantly closing under XWayland
On XWayland, the Wayland surface resource id is delivered via a ClientMessage
which does not need to be subscribed to.

Previously we were treating any ClientMessage as an instant 'close window' and
not checking it because we assumed that it was always that atom.

This was problematic as the XWayland surface resource id message would make
us trigger that instant close path.

Instead, change the code such that we check the associated atom of the client
message matches the delete message event.

Signed-off-by: Joshua Ashton <joshua@froggi.es>